Restructured components and changed imports.
