+++ src/components/Dashboard/Dashboard.tsx
@@ -2,8 +2,6 @@ interface DashboardProps {
   children?: React.ReactChildren;
 }
 
-const Dashboard = ({ children }: DashboardProps) => {
+export const Dashboard = ({ children }: DashboardProps) => {
   return <>{children}</>;
 };
-
-export default Dashboard;
