+++ src/components/SideMenu/SideMenuItem.tsx
@@ -1,0 +1,7 @@
+import Link from "next/link";
+import { Box } from "@chakra-ui/react";
+const SideMenuItem = () => {
+  return <Box>{/* <Link */}</Box>;
+};
+
+export default SideMenuItem;
