--- src/components/SideMenuItem.tsx
@@ -1,11 +1,0 @@
-import Link from 'next/link'
-import { Box } from '@chakra-ui/react'
-const NavItem = () => {
-  return (
-    <Box>
-      <Link
-    </Box>
-  )
-}
-
-export default NavItem
