+++ src/components/index.tsx
@@ -1,0 +1,7 @@
+export { Auth } from "src/components/Auth/Auth";
+export { DarkModeSwitch } from "src/components/DarkModeSwitch/DarkModeSwitch";
+export { Dashboard } from "src/components/Dashboard/Dashboard";
+export { Header } from "src/components/Header/Header";
+export { Loading } from "src/components/Loading/Loading";
+export { Page } from "src/components/Page/Page";
+export { SideMenu } from "src/components/SideMenu/SideMenu";
