+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import { SessionProvider } from "next-auth/react";
-import { Auth } from "../components/Auth";
+import { Auth } from "src/components";
 import theme from "../theme";
 import { AppProps } from "next/app";
 import "@fontsource/inter/400.css";
